Obliterate FileHandler
