--- src/main/java/com/samczsun/helios/handler/files/ClassHandler.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2016 Sam Sun <me@samczsun.com>
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-package com.samczsun.helios.handler.files;
-
-import org.eclipse.swt.custom.CTabFolder;
-import org.eclipse.swt.widgets.Control;
-
-public class ClassHandler extends FileHandler {
-    @Override
-    public Control generateTab(CTabFolder parent, String file, String name) {
-        return null;
-    }
-
-    @Override
-    public boolean accept(String name) {
-        return name.endsWith(".class");
-    }
-
-    @Override
-    public String getId() {
-        return "class";
-    }
-}
